hygiene: one walker, one quill builder — close #1062
§1 — the four hand-rolled `fs::read_dir` → FileTreeNode walkers are gone.
`quillmark::tree_from_path` (the tree half of `quill_from_path`, now public)
is the one walk, and it honours `.quillignore` and the default ignore set,
which none of the copies did. The Typst backend takes path-only dev-deps on
`quillmark` + `quillmark-fixtures`, the arrangement pdfform already
documents.

§2 — `tests/common/mod.rs` in the Typst backend carries `quill` (the wide
`files: &[(&str, &[u8])]` signature plate_resolution needs),
`quill_with_plate`, `host_tree` and `host_with_plate`. Six call sites, not
the five filed: sig_field.rs and live_apply.rs each inline a second copy.
`span_scan.rs`'s builder stays — a `src/` unit test cannot see `tests/`.

§3 — `decode_pdf_text`/`widget` to a pdfform `tests/common`;
`quill_from_yaml` to one `#[cfg(test)] pub(crate)` fn at the quill module
root; `collect_md_files` shared by the two files that walk recursively.
`emit_tests::fixtures_round_trip` keeps its own top-level scan — the walks
are not equivalent, the recursive one also picks up package README/LICENSE
files that parse-skip.

Not done, and why: the tempdir quill builders differ in five parameters
across three files, so a shared helper reads worse than the inline writes
(the issue's own suggestion there is to merge the files, which #1056 also
touches); the `content/` fixture builders the issue already calls marginal.
