+++ crates/backends/typst/Cargo.toml
@@ -38,3 +38,9 @@ wasm-bindgen = { workspace = true }
 tempfile = { workspace = true }
 # Acceptance tests parse the injected PDF to validate AcroForm structure.
 lopdf = { workspace = true }
+quillmark-fixtures = { workspace = true }
+# Path-only: `quillmark` depends on this crate, so a versioned dev-dep here is
+# an unpublishable cycle; cargo strips path-only dev-deps at packaging. Tests
+# load fixture quills through its `tree_from_path`, the one walker that honours
+# `.quillignore`.
+quillmark = { path = "../../quillmark" }
