+++ crates/backends/typst/tests/common/mod.rs
@@ -1,0 +1,60 @@
+//! Quill builders shared by the backend's acceptance tests.
+//!
+//! Two shapes: a self-contained quill assembled from literals, and the
+//! `usaf_memo` fixture used as a host when a test needs real fonts and
+//! packages but nothing quill-specific.
+
+// Each integration test binary compiles this module and uses part of it.
+#![allow(dead_code)]
+
+use quillmark_core::{FileTreeNode, Quill};
+use std::collections::HashMap;
+
+/// A self-contained quill from a `Quill.yaml` plus the named files. No fonts
+/// dir is needed — Typst's embedded defaults render text — and the helper
+/// package (`@local/quillmark-helper`) is injected by the backend.
+pub fn quill(yaml: &str, files: &[(&str, &[u8])]) -> Quill {
+    let mut map = HashMap::new();
+    map.insert(
+        "Quill.yaml".to_string(),
+        FileTreeNode::File {
+            contents: yaml.as_bytes().to_vec(),
+        },
+    );
+    for (name, bytes) in files {
+        map.insert(
+            (*name).to_string(),
+            FileTreeNode::File {
+                contents: bytes.to_vec(),
+            },
+        );
+    }
+    Quill::from_tree(FileTreeNode::Directory { files: map }).expect("load quill")
+}
+
+/// [`quill`] for the common `Quill.yaml` + `plate.typ` pair.
+pub fn quill_with_plate(yaml: &str, plate: &str) -> Quill {
+    quill(yaml, &[("plate.typ", plate.as_bytes())])
+}
+
+/// The `usaf_memo` fixture as an in-memory tree. Reused as a host because the
+/// features under test don't depend on any quill-specific config — any valid
+/// quill skeleton (fonts, packages) works.
+pub fn host_tree() -> FileTreeNode {
+    quillmark::tree_from_path(quillmark_fixtures::quills_path("usaf_memo")).expect("walk fixture")
+}
+
+/// The host quill with its `plate.typ` replaced by `plate`; the fixture's
+/// `typst.plate_file: plate.typ` makes the backend read this override.
+pub fn host_with_plate(plate: &str) -> Quill {
+    let mut tree = host_tree();
+    if let FileTreeNode::Directory { files } = &mut tree {
+        files.insert(
+            "plate.typ".to_string(),
+            FileTreeNode::File {
+                contents: plate.as_bytes().to_vec(),
+            },
+        );
+    }
+    Quill::from_tree(tree).expect("load source")
+}
